Ensure select row after re-populate works
In some cases, you could select a row, then re-populate the grid, but the
row remains in the new filter. If you then click on the row, the events
don't fire to load the row details as it thinks the row is already
loaded.

Fixes https://github.com/BiologicalRecordsCentre/iRecord/issues/1111